Iter 2

+++ src/ui/includes/input_field.hpp
@@ -5,9 +5,10 @@
 
 class InputField {
  public:
-  InputField(double x, double y, double width, double height, std::string_view text = "");
+  InputField(int x, int y, double width, double height,
+             std::string_view text = "");
 
-  void Draw(sf::RenderWindow& window);
+  void Draw(sf::RenderWindow &window);
   void HandleEvent(sf::Event event);
   std::string GetText() const;
